MR-51: Fixed all returns in the same file
