--- src/Components/Members/Form/form.module.css
@@ -1,44 +1,0 @@
-.form {
-    width: 100%;
-    max-width: 500px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: #f2f2f2;
-}
-
-.subContainer {
-    display: flex;
-    flex-direction: column;
-}
-
-.inputContainer {
-    margin-bottom: 10px;
-}
-
-.label {
-    font-size: 14px;
-    font-weight: bold;
-}
-
-.input {
-    padding: 5px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    width: 100%;
-    box-sizing: border-box;
-    font-size: 14px;
-}
-
-.button {
-    padding: 10px 20px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    background-color: #4caf50;
-    color: #fff;
-    font-size: 16px;
-}
-
-.button:hover {
-    background-color: #45a049;
-}
