--- src/Components/Members/Table/table.module.css
@@ -1,33 +1,0 @@
-.tableMember {
-    border-collapse: collapse;
-    width: 100%;
-}
-
-.thMember,
-.tdMember {
-    padding: 8px;
-    text-align: left;
-    border-bottom: 1px solid #ddd;
-}
-
-.thMember {
-    background-color: #f2f2f2;
-}
-
-.editButton {
-    background-color: #4caf50;
-    padding: 5px 10px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    color: #fff;
-}
-
-.deleteButton {
-    background-color: #f44336;
-    padding: 5px 10px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    color: #fff;
-}
