+++ src/Components/Members/index.jsx
@@ -1,7 +1,5 @@
 import { useEffect, useState } from 'react';
 import styles from './members.module.css';
-import Table from './Table';
-import Form from './Form';
 
 const Members = () => {
   const [members, setMembers] = useState([]);
@@ -29,6 +27,36 @@ const Members = () => {
   useEffect(() => {
     getMembers();
   }, []);
+
+  const onChange = (e) => {
+    setMemberValues({
+      ...memberValues,
+      [e.target.name]: e.target.value
+    });
+  };
+
+  const onSubmit = (e) => {
+    e.preventDefault();
+    if (idMember === '') {
+      addMember(memberValues);
+    } else {
+      updateMember(idMember);
+    }
+    setShowForm(false);
+    setIdMember('');
+    setMemberValues({
+      firstName: '',
+      lastName: '',
+      email: '',
+      dni: '',
+      phone: '',
+      city: '',
+      birthday: '2023-05-29T07:32:26+0000',
+      postalCode: '',
+      membership: '',
+      isActive: true
+    });
+  };
 
   const changeDateFormat = (date) => {
     let dateArray = date.split('-');
@@ -101,24 +129,204 @@ const Members = () => {
     <section className={styles.container}>
       <h2>Members</h2>
       {showForm ? (
-        <Form
-          addMember={addMember}
-          memberValues={memberValues}
-          setMemberValues={setMemberValues}
-          updateMember={updateMember}
-          idMember={idMember}
-          setIdMember={setIdMember}
-          setShowForm={setShowForm}
-        />
+        <form className={styles.form} onSubmit={onSubmit}>
+          <div className={styles.subContainer}>
+            <div className={styles.inputContainer}>
+              <label className={styles.label}>First Name</label>
+              <input
+                className={styles.input}
+                name="firstName"
+                type="text"
+                value={memberValues.firstName}
+                onChange={(e) => onChange(e)}
+              />
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>Last Name</label>
+                <input
+                  className={styles.input}
+                  name="lastName"
+                  type="text"
+                  value={memberValues.lastName}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>Email</label>
+                <input
+                  className={styles.input}
+                  name="email"
+                  type="text"
+                  value={memberValues.email}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>DNI</label>
+                <input
+                  className={styles.input}
+                  name="dni"
+                  type="number"
+                  value={memberValues.dni}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>Phone</label>
+                <input
+                  className={styles.input}
+                  name="phone"
+                  type="number"
+                  value={memberValues.phone}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>City</label>
+                <input
+                  className={styles.input}
+                  name="city"
+                  type="text"
+                  value={memberValues.city}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>Birthday</label>
+                <input
+                  className={styles.input}
+                  name="birthday"
+                  type="date"
+                  value={memberValues.birthday}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>PostalCode</label>
+                <input
+                  className={styles.input}
+                  name="postalCode"
+                  type="number"
+                  value={memberValues.postalCode}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>Membership</label>
+                <input
+                  className={styles.input}
+                  name="membership"
+                  type="text"
+                  value={memberValues.membership}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <div>
+              <div className={styles.inputContainer}>
+                <label className={styles.label}>IsActive</label>
+                <input
+                  className={styles.input}
+                  name="isActive"
+                  type="checkbox"
+                  value={memberValues.isActive}
+                  onChange={(e) => onChange(e)}
+                />
+              </div>
+            </div>
+            <button className={styles.button} type="submit">
+              Add
+            </button>
+            <button
+              className={styles.button}
+              onClick={() => {
+                setShowForm(false);
+                setIdMember('');
+                setMemberValues({
+                  firstName: '',
+                  lastName: '',
+                  email: '',
+                  dni: '',
+                  phone: '',
+                  city: '',
+                  birthday: '2023-05-29T07:32:26+0000',
+                  postalCode: '',
+                  membership: '',
+                  isActive: true
+                });
+              }}
+            >
+              Cancel
+            </button>
+          </div>
+        </form>
       ) : (
         <>
-          <Table
-            members={members}
-            deleteMember={deleteMember}
-            setIdMember={setIdMember}
-            setMemberValues={setMemberValues}
-            setShowForm={setShowForm}
-          />
+          <table className={styles.tableMember}>
+            <thead>
+              <tr>
+                <th className={styles.thMember}>Name</th>
+                <th className={styles.thMember}>Last Name</th>
+              </tr>
+            </thead>
+            <tbody>
+              {members?.length > 0 &&
+                members?.map((member) => {
+                  return (
+                    <tr key={member._id}>
+                      <td className={styles.tdMember}>{member.firstName}</td>
+                      <td className={styles.tdMember}>{member.lastName}</td>
+                      <td className={styles.tdMember}>
+                        <button
+                          className={styles.editButton}
+                          onClick={() => {
+                            setIdMember(member._id);
+                            setShowForm(true);
+                            setMemberValues({
+                              firstName: member.firstName,
+                              lastName: member.lastName,
+                              email: member.email,
+                              dni: member.dni,
+                              phone: member.phone,
+                              city: member.city,
+                              birthday: member.birthday,
+                              postalCode: member.postalCode,
+                              membership: member.membership,
+                              isActive: member.isActive
+                            });
+                          }}
+                        >
+                          Edit
+                        </button>
+                      </td>
+                      <td className={styles.tdMember}>
+                        <button
+                          className={styles.deleteButton}
+                          onClick={() => deleteMember(member._id)}
+                        >
+                          X
+                        </button>
+                      </td>
+                    </tr>
+                  );
+                })}
+            </tbody>
+          </table>
           <button className={styles.button} onClick={() => setShowForm(true)}>
             Add
           </button>
